backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
feedparser==6.0.10
newspaper3k==0.2.8
sentence-transformers==2.3.1
huggingface-hub==0.20.0
qdrant-client==1.7.0
redis==5.0.1
celery==5.3.4
spacy==3.7.2
openai==1.3.0
httpx==0.25.1
python-multipart==0.0.6